--- src/main/java/roomescape/auth/dto/LoginCheckResponse.java
@@ -1,4 +1,0 @@
-package roomescape.auth.dto;
-
-public record LoginCheckResponse(String name) {
-}
