+++ src/main/java/roomescape/auth/dto/response/CheckLoginResponse.java
@@ -1,0 +1,4 @@
+package roomescape.auth.dto.response;
+
+public record CheckLoginResponse(String name) {
+}
