+++ src/main/java/roomescape/auth/service/AuthService.java
@@ -3,8 +3,8 @@ package roomescape.auth.service;
 import org.springframework.stereotype.Service;
 
 import roomescape.auth.dto.LoginMember;
-import roomescape.auth.dto.LoginRequest;
-import roomescape.auth.dto.LoginResponse;
+import roomescape.auth.dto.request.LoginRequest;
+import roomescape.auth.dto.response.LoginResponse;
 import roomescape.common.exception.EntityNotFoundException;
 import roomescape.common.exception.LoginFailException;
 import roomescape.member.domain.Email;
