+++ src/main/java/roomescape/reservation/dto/request/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.dto;
+package roomescape.reservation.dto.request;
 
 import java.time.LocalDate;
