+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -19,12 +19,12 @@ import roomescape.member.repository.MemberRepository;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.Theme;
-import roomescape.reservation.dto.BookedReservationTimeResponse;
-import roomescape.reservation.dto.FilteringReservationRequest;
-import roomescape.reservation.dto.MyReservationsResponse;
-import roomescape.reservation.dto.ReservationCreateRequest;
-import roomescape.reservation.dto.ReservationResponse;
-import roomescape.reservation.dto.ReservationTimeResponse;
+import roomescape.reservation.dto.response.BookedReservationTimeResponse;
+import roomescape.reservation.dto.request.FilteringReservationRequest;
+import roomescape.reservation.dto.response.MyReservationsResponse;
+import roomescape.reservation.dto.request.ReservationCreateRequest;
+import roomescape.reservation.dto.response.ReservationResponse;
+import roomescape.reservation.dto.response.ReservationTimeResponse;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.repository.ReservationTimeRepository;
 import roomescape.reservation.repository.ThemeRepository;
